+++ pyannote/metrics/base.py
@@ -3,7 +3,7 @@
 
 # The MIT License (MIT)
 
-# Copyright (c) 2012-2019 CNRS
+# Copyright (c) 2012- CNRS
 
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -25,21 +25,20 @@
 
 # AUTHORS
 # Hervé BREDIN - http://herve.niderb.fr
+from typing import List, Union, Optional, Set, Tuple
 
-
-import scipy.stats
-import pandas as pd
+import warnings
 import numpy as np
+import pandas as pd
+import scipy.stats
+from pyannote.core import Annotation, Timeline
+
+from pyannote.metrics.types import Details, MetricComponents
 
 
-class BaseMetric(object):
+class BaseMetric:
     """
     :class:`BaseMetric` is the base class for most pyannote evaluation metrics.
-
-    Parameters
-    ----------
-    parallel : bool, optional
-        Defaults to True
 
     Attributes
     ----------
@@ -48,57 +47,58 @@ class BaseMetric(object):
     """
 
     @classmethod
-    def metric_name(cls):
+    def metric_name(cls) -> str:
         raise NotImplementedError(
             cls.__name__ + " is missing a 'metric_name' class method. "
-            "It should return the name of the metric as string.")
+                           "It should return the name of the metric as string."
+        )
 
     @classmethod
-    def metric_components(cls):
+    def metric_components(cls) -> MetricComponents:
         raise NotImplementedError(
             cls.__name__ + " is missing a 'metric_components' class method. "
-            "It should return the list of names of metric components.")
+                           "It should return the list of names of metric components."
+        )
 
-    def __init__(self, parallel=False, **kwargs):
+    def __init__(self, **kwargs):
         super(BaseMetric, self).__init__()
-        self.parallel = parallel
         self.metric_name_ = self.__class__.metric_name()
-        self.components_ = set(self.__class__.metric_components())
+        self.components_: Set[str] = set(self.__class__.metric_components())
         self.reset()
 
     def init_components(self):
-        return {value: 0. for value in self.components_}
+        return {value: 0.0 for value in self.components_}
 
     def reset(self):
         """Reset accumulated components and metric values"""
-        if self.parallel:
-            from pyannote.metrics import manager_
-            self.accumulated_ = manager_.dict()
-            self.results_ = manager_.list()
-            self.uris_ = manager_.dict()
-        else:
-            self.accumulated_ = dict()
-            self.results_ = list()
-            self.uris_ = dict()
+        self.accumulated_: Details = dict()
+        self.results_: List = list()
         for value in self.components_:
-            self.accumulated_[value] = 0.
+            self.accumulated_[value] = 0.0
 
-    def __get_name(self):
-        return self.__class__.metric_name()
-    name = property(fget=__get_name, doc="Metric name.")
+    @property
+    def name(self):
+        """Metric name."""
+        return self.metric_name()
 
-    def __call__(self, reference, hypothesis, detailed=False, **kwargs):
+    # TODO: use joblib/locky to allow parallel processing?
+    # TODO: signature could be something like __call__(self, reference_iterator, hypothesis_iterator, ...)
+
+    def __call__(self, reference: Union[Timeline, Annotation],
+                 hypothesis: Union[Timeline, Annotation],
+                 detailed: bool = False, uri: Optional[str] = None, **kwargs):
         """Compute metric value and accumulate components
 
         Parameters
         ----------
         reference : type depends on the metric
             Manual `reference`
-        hypothesis : same as `reference`
+        hypothesis : type depends on the metric
             Evaluated `hypothesis`
+        uri : optional
+            Override uri.
         detailed : bool, optional
             By default (False), return metric value only.
-
             Set `detailed` to True to return dictionary where keys are
             components names and values are component values
 
@@ -108,7 +108,6 @@ class BaseMetric(object):
             Metric value
         components : dict (if `detailed` is True)
             `components` updated with metric value
-
         """
 
         # compute metric components
@@ -118,15 +117,7 @@ class BaseMetric(object):
         components[self.metric_name_] = self.compute_metric(components)
 
         # keep track of this computation
-        uri = reference.uri
-        if uri is None:
-            uri = "???"
-        if uri not in self.uris_:
-            self.uris_[uri] = 1
-        else:
-            self.uris_[uri] += 1
-            uri = uri + ' #{0:d}'.format(self.uris_[uri])
-
+        uri = uri or getattr(reference, "uri", "NA")
         self.results_.append((uri, components))
 
         # accumulate components
@@ -138,7 +129,7 @@ class BaseMetric(object):
 
         return components[self.metric_name_]
 
-    def report(self, display=False):
+    def report(self, display: bool = False) -> pd.DataFrame:
         """Evaluation report
 
         Parameters
@@ -156,24 +147,24 @@ class BaseMetric(object):
         report = []
         uris = []
 
-        percent = 'total' in self.metric_components()
+        percent = "total" in self.metric_components()
 
         for uri, components in self.results_:
             row = {}
             if percent:
-                total = components['total']
+                total = components["total"]
             for key, value in components.items():
                 if key == self.name:
-                    row[key, '%'] = 100 * value
-                elif key == 'total':
-                    row[key, ''] = value
+                    row[key, "%"] = 100 * value
+                elif key == "total":
+                    row[key, ""] = value
                 else:
-                    row[key, ''] = value
+                    row[key, ""] = value
                     if percent:
                         if total > 0:
-                            row[key, '%'] = 100 * value / total
+                            row[key, "%"] = 100 * value / total
                         else:
-                            row[key, '%'] = np.NaN
+                            row[key, "%"] = np.NaN
 
             report.append(row)
             uris.append(uri)
@@ -182,50 +173,57 @@ class BaseMetric(object):
         components = self.accumulated_
 
         if percent:
-            total = components['total']
+            total = components["total"]
 
         for key, value in components.items():
             if key == self.name:
-                row[key, '%'] = 100 * value
-            elif key == 'total':
-                row[key, ''] = value
+                row[key, "%"] = 100 * value
+            elif key == "total":
+                row[key, ""] = value
             else:
-                row[key, ''] = value
+                row[key, ""] = value
                 if percent:
                     if total > 0:
-                        row[key, '%'] = 100 * value / total
+                        row[key, "%"] = 100 * value / total
                     else:
-                        row[key, '%'] = np.NaN
+                        row[key, "%"] = np.NaN
 
-        row[self.name, '%'] = 100 * abs(self)
+        row[self.name, "%"] = 100 * abs(self)
         report.append(row)
-        uris.append('TOTAL')
+        uris.append("TOTAL")
 
         df = pd.DataFrame(report)
 
-        df['item'] = uris
-        df = df.set_index('item')
+        df["item"] = uris
+        df = df.set_index("item")
 
         df.columns = pd.MultiIndex.from_tuples(df.columns)
 
         df = df[[self.name] + self.metric_components()]
 
         if display:
-            print(df.to_string(index=True, sparsify=False, justify='right', float_format=lambda f: '{0:.2f}'.format(f)))
+            print(
+                df.to_string(
+                    index=True,
+                    sparsify=False,
+                    justify="right",
+                    float_format=lambda f: "{0:.2f}".format(f),
+                )
+            )
 
         return df
 
     def __str__(self):
         report = self.report(display=False)
         return report.to_string(
-            sparsify=False,
-            float_format=lambda f: '{0:.2f}'.format(f))
+            sparsify=False, float_format=lambda f: "{0:.2f}".format(f)
+        )
 
     def __abs__(self):
         """Compute metric value from accumulated components"""
         return self.compute_metric(self.accumulated_)
 
-    def __getitem__(self, component):
+    def __getitem__(self, component: str) -> Union[float, Details]:
         """Get value of accumulated `component`.
 
         Parameters
@@ -249,7 +247,10 @@ class BaseMetric(object):
         for uri, component in self.results_:
             yield uri, component
 
-    def compute_components(self, reference, hypothesis, **kwargs):
+    def compute_components(self,
+                           reference: Union[Timeline, Annotation],
+                           hypothesis: Union[Timeline, Annotation],
+                           **kwargs) -> Details:
         """Compute metric components
 
         Parameters
@@ -268,10 +269,11 @@ class BaseMetric(object):
         """
         raise NotImplementedError(
             self.__class__.__name__ + " is missing a 'compute_components' method."
-            "It should return a dictionary where keys are component names "
-            "and values are component values.")
+                                      "It should return a dictionary where keys are component names "
+                                      "and values are component values."
+        )
 
-    def compute_metric(self, components):
+    def compute_metric(self, components: Details):
         """Compute metric value from computed `components`
 
         Parameters
@@ -287,10 +289,12 @@ class BaseMetric(object):
         """
         raise NotImplementedError(
             self.__class__.__name__ + " is missing a 'compute_metric' method. "
-            "It should return the actual value of the metric based "
-            "on the precomputed component dictionary given as input.")
+                                      "It should return the actual value of the metric based "
+                                      "on the precomputed component dictionary given as input."
+        )
 
-    def confidence_interval(self, alpha=0.9):
+    def confidence_interval(self, alpha: float = 0.9) \
+            -> Tuple[float, Tuple[float, float]]:
         """Compute confidence interval on accumulated metric values
 
         Parameters
@@ -311,14 +315,24 @@ class BaseMetric(object):
         scipy.stats.bayes_mvs
 
         """
-        m, _, _ = scipy.stats.bayes_mvs(
-            [r[self.metric_name_] for _, r in self.results_], alpha=alpha)
-        return m
+
+        values = [r[self.metric_name_] for _, r in self.results_]
+
+        if len(values) == 0:
+            raise ValueError("Please evaluate a bunch of files before computing confidence interval.")
+        
+        elif len(values) == 1:
+            warnings.warn("Cannot compute a reliable confidence interval out of just one file.")
+            center = lower = upper = values[0]
+            return center, (lower, upper)
+        
+        else:
+            return scipy.stats.bayes_mvs(values, alpha=alpha)[0]
 
 
-PRECISION_NAME = 'precision'
-PRECISION_RETRIEVED = '# retrieved'
-PRECISION_RELEVANT_RETRIEVED = '# relevant retrieved'
+PRECISION_NAME = "precision"
+PRECISION_RETRIEVED = "# retrieved"
+PRECISION_RELEVANT_RETRIEVED = "# relevant retrieved"
 
 
 class Precision(BaseMetric):
@@ -338,25 +352,25 @@ class Precision(BaseMetric):
         return PRECISION_NAME
 
     @classmethod
-    def metric_components(cls):
+    def metric_components(cls) -> MetricComponents:
         return [PRECISION_RETRIEVED, PRECISION_RELEVANT_RETRIEVED]
 
-    def compute_metric(self, components):
+    def compute_metric(self, components: Details) -> float:
         """Compute precision from `components`"""
         numerator = components[PRECISION_RELEVANT_RETRIEVED]
         denominator = components[PRECISION_RETRIEVED]
-        if denominator == 0.:
+        if denominator == 0.0:
             if numerator == 0:
-                return 1.
+                return 1.0
             else:
-                raise ValueError('')
+                raise ValueError("")
         else:
-            return numerator/denominator
+            return numerator / denominator
 
 
-RECALL_NAME = 'recall'
-RECALL_RELEVANT = '# relevant'
-RECALL_RELEVANT_RETRIEVED = '# relevant retrieved'
+RECALL_NAME = "recall"
+RECALL_RELEVANT = "# relevant"
+RECALL_RELEVANT_RETRIEVED = "# relevant retrieved"
 
 
 class Recall(BaseMetric):
@@ -376,23 +390,23 @@ class Recall(BaseMetric):
         return RECALL_NAME
 
     @classmethod
-    def metric_components(cls):
+    def metric_components(cls) -> MetricComponents:
         return [RECALL_RELEVANT, RECALL_RELEVANT_RETRIEVED]
 
-    def compute_metric(self, components):
+    def compute_metric(self, components: Details) -> float:
         """Compute recall from `components`"""
         numerator = components[RECALL_RELEVANT_RETRIEVED]
         denominator = components[RECALL_RELEVANT]
-        if denominator == 0.:
+        if denominator == 0.0:
             if numerator == 0:
-                return 1.
+                return 1.0
             else:
-                raise ValueError('')
+                raise ValueError("")
         else:
-            return numerator/denominator
+            return numerator / denominator
 
 
-def f_measure(precision, recall, beta=1.):
+def f_measure(precision: float, recall: float, beta=1.0) -> float:
     """Compute f-measure
 
     f-measure is defined as follows:
@@ -402,4 +416,4 @@ def f_measure(precision, recall, beta=1.):
     """
     if precision + recall == 0.0:
         return 0
-    return (1+beta*beta)*precision*recall / (beta*beta*precision+recall)
+    return (1 + beta * beta) * precision * recall / (beta * beta * precision + recall)